Replaced Node.js specific code
